Adding modal component
