+++ src/App.js
@@ -1,26 +1,13 @@
 import React, { Component } from 'react';
 import './App.css';
-import { FantasyFootballApi } from './FantasyFootballApi';
+import { FantasyFootballApi } from './api/FantasyFootballApi';
+import UrlEntryModal from './components/UrlEntryModal';
 import { Segment, Table, Header, Image, Grid, Container, Form, Button, Loader, Modal, Icon, Message } from 'semantic-ui-react';
 import '../node_modules/semantic-ui-css/semantic.min.css';
-
-const getParams = query => {
-  if (!query) {
-    return { };
-  }
-
-  return (query
-    .split('?')[1]
-    .split('&')
-    .reduce((params, param) => {
-      let [ key, value ] = param.split('=');
-      params[key] = value ? decodeURIComponent(value.replace(/\+/g, ' ')) : '';
-      return params;
-    }, { }));
-};
+import { getParams } from './util/utils'
 
 class App extends Component {
-  
+
   constructor() {
     super();
     this.state = {
@@ -33,8 +20,7 @@ class App extends Component {
     this.api = new FantasyFootballApi();
   }
 
-  async onSubmit() {
-    let params = getParams(this.state.leagueUrl);
+  async onSubmit(params) {
     this.setState({rankings: [], loading: true, modalOpen: false})
     try {
       let rankings = await this.api.getPowerRankings(params.leagueId, params.seasonId);
@@ -50,10 +36,6 @@ class App extends Component {
     }
   }
 
-  handleLeagueUrlChange = e => {
-    this.setState({leagueUrl: e.target.value})
-  }
-
   handleOpen = () => this.setState({ modalOpen: true })
 
   handleClose = () => this.setState({ modalOpen: false })
@@ -61,33 +43,12 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        <Modal
-          open={this.state.modalOpen}
-          onClose={this.handleClose}
-          closeOnDimmerClick={false}
-          basic
-          size='tiny'
-        >
-          <Header icon='write' content='Enter ESPN League URL' />
-          <Modal.Content>
-            {this.state.error && <Message negative>
-              <Message.Header>We're sorry, something went wrong. Please try again.</Message.Header>
-            </Message>}
-            <Form onSubmit={this.onSubmit.bind(this)}>
-              <Form.Field>
-                <input 
-                  placeholder='http://games.espn.com/ffl/leagueoffice?leagueId=123456&seasonId=2017' 
-                  value={this.state.leagueUrl} 
-                  onChange={this.handleLeagueUrlChange}/>
-              </Form.Field>
-            </Form>
-          </Modal.Content>
-          <Modal.Actions>
-            <Button color='green' type='submit' onClick={this.onSubmit.bind(this)} inverted>
-              <Icon name='checkmark' /> Submit
-            </Button>
-          </Modal.Actions>
-        </Modal>
+        <UrlEntryModal
+          onSubmit={this.onSubmit.bind(this)}
+          modalOpen={this.state.modalOpen}
+          handleClose={this.handleClose.bind(this)}
+          error={this.state.error}
+        />
         <Loader active={this.state.loading}/>
         <Container>
         <Grid centered>
